refactor: subaccount file names match other modules
