+++ proto/sge/subaccount/query.proto
@@ -23,20 +23,15 @@ service Query {
 // QueryParamsRequest is the request type for the Query/Params RPC method
 message QueryParamsRequest {}
 // QueryParamsResponse is the response type for the Query/Params RPC method
-message QueryParamsResponse {
-  sge.subaccount.Params params = 1;
-}
+message QueryParamsResponse { sge.subaccount.Params params = 1; }
 
 // QuerySubaccountRequest is the request type for the Query/Subaccount RPC
-message QuerySubaccountRequest {
-  string subaccount_owner = 1;
-}
+message QuerySubaccountRequest { string subaccount_owner = 1; }
 
 // QuerySubaccountResponse is the response type for the Query/Subaccount RPC
 message QuerySubaccountResponse {
   string subaccount_address = 1;
-  sge.subaccount.Balance balance = 2 [(gogoproto.nullable) = false ];
-  repeated sge.subaccount.LockedBalance locked_balance = 3 [(gogoproto.nullable) = false ];
+  sge.subaccount.Balance balance = 2 [ (gogoproto.nullable) = false ];
+  repeated sge.subaccount.LockedBalance locked_balance = 3
+      [ (gogoproto.nullable) = false ];
 }
-
-
