+++ x/subaccount/keeper/msg_server_bet_test.go
@@ -60,7 +60,7 @@ func TestMsgServer_Bet(t *testing.T) {
 		),
 	)
 
-	_, err := msgServer.CreateSubAccount(sdk.WrapSDKContext(ctx), &types.MsgCreateSubAccount{
+	_, err := msgServer.Create(sdk.WrapSDKContext(ctx), &types.MsgCreate{
 		Sender:          subAccFunder.String(),
 		SubAccountOwner: subAccOwner.String(),
 		LockedBalances: []types.LockedBalance{
