+++ x/subaccount/keeper/msg_server_house_test.go
@@ -37,7 +37,7 @@ func TestMsgServer(t *testing.T) {
 		),
 	)
 
-	_, err := msgServer.CreateSubAccount(sdk.WrapSDKContext(ctx), &types.MsgCreateSubAccount{
+	_, err := msgServer.Create(sdk.WrapSDKContext(ctx), &types.MsgCreate{
 		Sender:          subAccFunder.String(),
 		SubAccountOwner: subAccOwner.String(),
 		LockedBalances: []types.LockedBalance{
@@ -210,7 +210,7 @@ func TestHouseWithdrawal_MarketRefund(t *testing.T) {
 		),
 	)
 
-	_, err := msgServer.CreateSubAccount(sdk.WrapSDKContext(ctx), &types.MsgCreateSubAccount{
+	_, err := msgServer.Create(sdk.WrapSDKContext(ctx), &types.MsgCreate{
 		Sender:          subAccFunder.String(),
 		SubAccountOwner: subAccOwner.String(),
 		LockedBalances: []types.LockedBalance{
