+++ x/subaccount/keeper/query_server_test.go
@@ -30,7 +30,7 @@ func TestQueryServer(t *testing.T) {
 		),
 	)
 
-	_, err := msgServer.CreateSubAccount(sdk.WrapSDKContext(ctx), &types.MsgCreateSubAccount{
+	_, err := msgServer.Create(sdk.WrapSDKContext(ctx), &types.MsgCreate{
 		Sender:          subAccFunder.String(),
 		SubAccountOwner: subAccOwner.String(),
 		LockedBalances: []types.LockedBalance{
