+++ x/subaccount/types/messages_balance.go
@@ -1,0 +1,56 @@
+package types
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/errors"
+)
+
+var (
+	_ sdk.Msg = &MsgTopUp{}
+	_ sdk.Msg = &MsgWithdrawUnlockedBalances{}
+)
+
+func (msg *MsgTopUp) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		return errors.ErrInvalidAddress
+	}
+
+	_, err = sdk.AccAddressFromBech32(msg.SubAccount)
+	if err != nil {
+		return errors.ErrInvalidAddress
+	}
+
+	for _, balanceUnlock := range msg.LockedBalances {
+		if err = balanceUnlock.Validate(); err != nil {
+			return errors.Wrapf(err, "invalid locked balance")
+		}
+	}
+
+	return nil
+}
+
+func (msg *MsgTopUp) GetSigners() []sdk.AccAddress {
+	signer, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		panic(err)
+	}
+	return []sdk.AccAddress{signer}
+}
+
+func (msg *MsgWithdrawUnlockedBalances) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		return errors.ErrInvalidAddress
+	}
+
+	return nil
+}
+
+func (msg *MsgWithdrawUnlockedBalances) GetSigners() []sdk.AccAddress {
+	signer, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		panic(err)
+	}
+	return []sdk.AccAddress{signer}
+}
