+++ x/subaccount/types/messages_subaccount.go
@@ -1,0 +1,40 @@
+package types
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/errors"
+)
+
+var (
+	_ sdk.Msg = &MsgCreate{}
+)
+
+func (msg *MsgCreate) GetSigners() []sdk.AccAddress {
+	signer, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		panic(err)
+	}
+	return []sdk.AccAddress{signer}
+}
+
+// ValidateBasic performs a basic validation of the MsgCreateRequest fields.
+func (msg *MsgCreate) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(msg.Sender)
+	if err != nil {
+		return errors.ErrInvalidAddress
+	}
+
+	_, err = sdk.AccAddressFromBech32(msg.SubAccountOwner)
+
+	if err != nil {
+		return errors.ErrInvalidAddress
+	}
+
+	for _, balanceUnlock := range msg.LockedBalances {
+		if err = balanceUnlock.Validate(); err != nil {
+			return errors.Wrapf(err, "invalid locked balance")
+		}
+	}
+
+	return nil
+}
